+++ src/server/controllers/continue.controller.test.js
@@ -6,7 +6,8 @@ const { moveAlongPath, editPath } = require("../services/path.service");
 const { validate } = require("../services/validation.service");
 const {
   cleanInactivePathAnswers,
-  cleanEmptiedAnswers
+  cleanEmptiedAnswers,
+  cleanSwitches
 } = require("../services/session-management.service");
 
 const continueController = require("./continue.controller");
@@ -18,6 +19,7 @@ describe("Function: continueController: ", () => {
       (previousAnswers, newAnswersArray) =>
         newAnswersArray.length > 0 ? previousAnswers : null
     );
+    cleanSwitches.mockImplementation(() => ({}));
     editPath.mockImplementation(() => ({
       "/some-page": {
         on: true,
@@ -32,6 +34,8 @@ describe("Function: continueController: ", () => {
 
   let response;
 
+  const exampleSwitches = { switch1: true, switch2: false };
+
   const exampleAnswers = {
     answer: "answer-pathAnswer"
   };
@@ -41,7 +45,12 @@ describe("Function: continueController: ", () => {
       validate.mockImplementation(() => ({
         errors: {}
       }));
-      response = continueController("/some-page", exampleAnswers, {});
+      response = continueController(
+        "/some-page",
+        exampleAnswers,
+        {},
+        exampleSwitches
+      );
     });
 
     it("Should return the same answers as input", () => {
@@ -59,7 +68,12 @@ describe("Function: continueController: ", () => {
         validate.mockImplementation(() => ({
           errors: {}
         }));
-        response = continueController("/final-page", {}, exampleAnswers);
+        response = continueController(
+          "/final-page",
+          {},
+          exampleAnswers,
+          exampleSwitches
+        );
       });
 
       it("Should set redirect route to /submit", () => {
@@ -73,7 +87,12 @@ describe("Function: continueController: ", () => {
           errors: {}
         }));
         moveAlongPath.mockImplementation(() => "/nextPage");
-        response = continueController("/some-page", {}, exampleAnswers);
+        response = continueController(
+          "/some-page",
+          {},
+          exampleAnswers,
+          exampleSwitches
+        );
       });
 
       it("Should return a controllerResponse", () => {
@@ -97,10 +116,107 @@ describe("Function: continueController: ", () => {
       validate.mockImplementation(() => ({
         errors: { some: "error" }
       }));
-      response = continueController("/mock-page-1", {}, exampleAnswers);
+      response = continueController(
+        "/mock-page-1",
+        {},
+        exampleAnswers,
+        exampleSwitches
+      );
     });
     it("should set redirectRoute to the currentPage", () => {
       expect(response.redirectRoute).toBe("/mock-page-1");
+    });
+  });
+
+  describe("When there are no switches: ", () => {
+    cleanSwitches.mockImplementation(() => {});
+
+    it("should return an empty switches object", () => {
+      const exampleEmptySwitches = [{}, undefined, null];
+
+      response = continueController(
+        "/mock-page-1",
+        {},
+        exampleAnswers,
+        exampleEmptySwitches
+      );
+
+      expect(response.switches).toEqual({});
+    });
+  });
+
+  describe("When switches are passed in: ", () => {
+    beforeEach(() => {
+      cleanSwitches.mockImplementation(() => exampleSwitches);
+    });
+
+    it("should return the same object keys", () => {
+      response = continueController(
+        "/mock-page-1",
+        {},
+        exampleAnswers,
+        exampleSwitches
+      );
+
+      const originalSwitchesKeyArray = Object.keys(exampleSwitches);
+      const responseSwitchesKeyArray = Object.keys(response.switches);
+
+      expect(originalSwitchesKeyArray).toEqual(responseSwitchesKeyArray);
+    });
+
+    it("should return boolean values", () => {
+      response = continueController(
+        "/mock-page-1",
+        {},
+        exampleAnswers,
+        exampleSwitches
+      );
+
+      const responseSwitchesValueArray = Object.values(response.switches);
+
+      responseSwitchesValueArray.forEach(value => {
+        expect(typeof value).toBe("boolean");
+      });
+    });
+
+    describe("when cleanSwitches changes the switches that were passed in", () => {
+      beforeEach(() => {
+        cleanSwitches.mockImplementation(() => ({
+          switch1: false,
+          switch2: true
+        }));
+      });
+
+      it("should return the result of cleanSwitches, not the original values", () => {
+        response = continueController(
+          "/mock-page-1",
+          {},
+          exampleAnswers,
+          exampleSwitches
+        );
+
+        expect(response.switches).toEqual({ switch1: false, switch2: true });
+      });
+    });
+
+    describe("when there are switches and validator errors", () => {
+      beforeEach(() => {
+        cleanSwitches.mockImplementation(() => ({
+          switch1: false,
+          switch2: true
+        }));
+      });
+
+      it("should process the switches as usual", () => {
+        response = continueController(
+          "/mock-page-1",
+          { errorName: "error" },
+          exampleAnswers,
+          exampleSwitches
+        );
+
+        expect(response.switches).toEqual({ switch1: false, switch2: true });
+      });
     });
   });
 });