+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## v1.15.0 - 10th July 2018 (unreleased)
+
+### Added
+
+* SDB-113 Added establishment contact details page with 'reuse operator contact details' button
+* Add switches route and controller
+
 ## v1.14.4 - 4th July 2018 (unreleased)
 
 ### Modified